+++ include/webcraft/async/io/fs.hpp
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "core.hpp"
+
+namespace webcraft::async::io::fs
+{
+
+}
