+++ include/webcraft/async/io/io.hpp
@@ -1,0 +1,6 @@
+#pragma once
+
+#include "core.hpp"
+#include "adaptors.hpp"
+#include "fs.hpp"
+#include "socket.hpp"
